refactor Service Controller to it's own file (#6395)
